+++ decadriver/deca_param_types.h
@@ -48,7 +48,7 @@ extern const uint32 tx_config[NUM_CH];
 extern const uint8 dwnsSFDlen[NUM_BR]; //length of SFD for each of the bitrates
 extern const uint32 digital_bb_config[NUM_PRF][NUM_PACS];
 extern const uint8 chan_idx[NUM_CH_SUPPORTED];
-extern const double txpwr_compensation[NUM_CH];
+extern const float txpwr_compensation[NUM_CH];
 
 #define PEAK_MULTPLIER  (0x60) //3 -> (0x3 * 32) & 0x00E0
 #define N_STD_FACTOR    (13)
